Fix detekt issues
