+++ orbit-2-core/src/main/java/com/babylon/orbit2/Container.kt
@@ -16,8 +16,6 @@
 
 package com.babylon.orbit2
 
-import kotlinx.coroutines.CoroutineScope
-
 /**
  * The heart of the Orbit MVI system. Represents an MVI container with its input and outputs.
  * You can manipulate the container through the [orbit] function
@@ -55,10 +53,6 @@ interface Container<STATE : Any, SIDE_EFFECT : Any> {
         init: Builder<STATE, SIDE_EFFECT, Unit>.() -> Builder<STATE, SIDE_EFFECT, *>
     )
 
-    @Suppress("EXPERIMENTAL_API_USAGE")
-    companion object {
-    }
-
     /**
      * Represents additional settings to create the container with.
      *
@@ -70,34 +64,3 @@ interface Container<STATE : Any, SIDE_EFFECT : Any> {
         val sideEffectCaching: Boolean = true
     )
 }
-
-/**
- * Helps create a concrete container in a standard way.
- *
- * @param initialState The initial state of the container.
- * @param settings The [Settings] to set the container up with.
- * @param onCreate The lambda to execute when the container is created. By default it is
- * executed in a lazy manner when the container is first interacted with in any way.
- * @return A [Container] implementation
- */
-fun <STATE : Any, SIDE_EFFECT : Any> CoroutineScope.container(
-    initialState: STATE,
-    settings: Container.Settings = Container.Settings(),
-    onCreate: (() -> Unit)? = null
-): Container<STATE, SIDE_EFFECT> =
-    if (onCreate == null) {
-        RealContainer(
-            initialState = initialState,
-            settings = settings,
-            parentScope = this
-        )
-    } else {
-        LazyCreateContainerDecorator(
-            RealContainer(
-                initialState = initialState,
-                settings = settings,
-                parentScope = this
-            ),
-            onCreate
-        )
-    }
